+++ changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## Upcoming
+### Additions
+* Config option to always eat, ignoring the full hunger bar
+
 ### Changes
 * Worlds should load faster as data packs are no longer loaded twice
 * Magic damage now bypasses again invincibility frames